Save list type arguments correctly
model is derived from type_ and we can't tell the complete type of this
argument from the model alone. Without this the detail page of an event
instance with a list argument will fail, since the argument has a list
of pks where the code expects a single pk for non-list arguments.